Split modelchecker::DFTModularizer into modelchecker::DftModularizationChecker and utility::DftModularizer
