+++ src/storm-dft/utility/MTTFHelper.cpp
@@ -1,5 +1,5 @@
 #include "storm-dft/utility/MTTFHelper.h"
-#include "storm-dft/modelchecker/DFTModularizer.h"
+#include "storm-dft/modelchecker/DftModularizationChecker.h"
 
 namespace {
 
@@ -23,7 +23,7 @@ namespace utility {
 
 double MTTFHelperProceeding(std::shared_ptr<storm::dft::storage::DFT<double>> const dft, double const stepsize, double const precision) {
     constexpr size_t chunksize{1001};
-    storm::dft::modelchecker::DFTModularizer checker{dft};
+    storm::dft::modelchecker::DftModularizationChecker checker{dft};
 
     std::vector<double> timepoints{};
     timepoints.resize(chunksize);
@@ -64,7 +64,7 @@ double MTTFHelperProceeding(std::shared_ptr<storm::dft::storage::DFT<double>> co
 
 double MTTFHelperVariableChange(std::shared_ptr<storm::dft::storage::DFT<double>> const dft, double const stepsize) {
     constexpr size_t chunksize{1001};
-    storm::dft::modelchecker::DFTModularizer checker{dft};
+    storm::dft::modelchecker::DftModularizationChecker checker{dft};
 
     std::vector<double> timepoints{};
     timepoints.resize(static_cast<size_t>(1 / stepsize) - 1);
